fix: changed dynamic node code
