--- packages/runtime/browser-vm/src/config.ts
@@ -1,11 +1,0 @@
-export const defaultConfig = {
-  modules: [],
-  baseUrl: '',
-  namespace: '',
-  useStrict: false,
-  openSandbox: true,
-  strictIsolation: false,
-  el: () => {},
-  protectVariable: () => [],
-  insulationVariable: () => [],
-};
