+++ packages/runtime/browser-vm/src/global.ts
@@ -7,24 +7,3 @@ declare module '@garfish/core' {
     }
   }
 }
-
-export let currentRunningSandbox: Sandbox | null = null;
-
-const sandboxMap = new WeakMap();
-
-export const setElementSandbox = function (element: Element, sandbox: Sandbox) {
-  if (sandboxMap.get(element)) return;
-  sandboxMap.set(element, sandbox);
-};
-
-export const getElementSandbox = function (element: Element): Sandbox {
-  return sandboxMap.get(element);
-};
-
-export const getCurrentRunningSandbox = function () {
-  return currentRunningSandbox;
-};
-
-export const setCurrentRunningSandbox = function (sandbox) {
-  currentRunningSandbox = sandbox;
-};
